Add files via upload
Given a training set of size 𝑁, we test the performance of a model trained on a subsample of size 𝑁𝑖, where 𝑁𝑖<=𝑁. We can plot how performance grows as we move 𝑁𝑖 from 0 to 𝑁

.

Because of the inherent randomness of subsamples of size 𝑁 𝑖, we should expect that any single sample of size 𝑁𝑖 might not be representative of an algorithm’s performance at a given training set size. To quantify this variance and get a better generalization, we will also use bootstrap analysis. In a bootstrap analysis, we pull multiple samples of size 𝑁𝑖, build a model, evaluate on a test set, and then take an average and standard error of the results.